Adds functioning dummy synthesizer 'hop'
