+++ examples/README.md
@@ -1,0 +1,4 @@
+Overview:
+
++ `example_mnist_cpu.py`
++ `example_mnist_gpu.py`
