+++ pyproject.toml
@@ -17,11 +17,11 @@ matplotlib = "^3.2.1"
 pytorch_lightning = "^0.7.6"
 ipykernel = "^5.3.0"
 attrs = "^19.3.0"
-adversarial-robustness-toolbox = "^1.3.3"
 sklearn = "^0.0"
 pytest = "^6.0.2"
 flake8 = "^3.8.3"
 isort = "^5.5.3"
+adversarial-robustness-toolbox = "^1.4.1"
 
 [tool.poetry.dev-dependencies]
 pytest = "^6.0"
